Fix build issues for math expressions and virtual workspace
