+++ src/virtual_lab/MathExpression.h
@@ -8,6 +8,8 @@
 
 namespace virtual_lab {
 
+class MathExpressionParser;
+
 class MathExpression {
  public:
   MathExpression();
@@ -22,7 +24,16 @@ class MathExpression {
   const String &expression() const { return expression_; }
 
  private:
-  struct Node;
+  friend class MathExpressionParser;
+
+  struct Node {
+    enum class Type { Constant, Variable, Unary, Binary, Function };
+    Type type;
+    float constant = 0.0f;
+    char op = '\0';
+    String identifier;
+    std::vector<std::unique_ptr<Node>> children;
+  };
 
   std::unique_ptr<Node> parseExpression(String &errorMessage);
